move additional examples with JSON serialization into main Client class
